backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
discord.py==2.4.0
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic==2.9.0

